requirements.txt
clarifai
streamlit
langchain
streamlit-card
moviepy==2.0.0.dev2
imageio==2.25.1
numpy
pandas
matplotlib
pillow
pydub
scipy
youtube-transcript-api
langchain
serpapi
python-dotenv

